+++ src/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { ref } from 'vue'
+import { ref, watch } from 'vue'
 import { GlobalSiteOptions } from './shared'
-import { GetSiteOptions } from './storage'
+import { GetSiteOptions, SaveSiteOptions } from './storage'
 
 import Chart from '#components/Chart.vue'
 import Sidebar from '#components/Sidebar.vue'
@@ -9,6 +9,24 @@ import Sidebar from '#components/Sidebar.vue'
 const renderChart = ref(false)
 
 GlobalSiteOptions.value = GetSiteOptions()
+
+// TODO: look into feasibility of debouncing this as to reduce ops
+watch(
+	GlobalSiteOptions,
+	() => {
+		if (!GlobalSiteOptions.value) {
+			return console.error(
+				'Error getting GlobalChartState in watch(GlobalChartState)',
+				GlobalSiteOptions.value
+			)
+		}
+
+		SaveSiteOptions()
+	},
+	{
+		deep: true
+	}
+)
 </script>
 
 <template>