+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 
 ### User facing
 
-- Feature descriptions - some explanation of some less obvious functionality
-- Chart Text: Font, size, and spacing options
-- Chart Padding adjustment
+- Chart Text: Font Selector
+- Chart Text: size and spacing sliders
+- Chart Padding slider
 
 ### Dev Specific
 
 - Updated PR Check to ensure Changelog is updated and package version is updated
 - Enhanced fetch() wrapper
 - Condense Tailwind to have more custom classes to K.I.S.S
+- Major refactor of Options component to not be a monolith